(refactor) components exporting / importing
