+++ app/src/components/main/index.tsx
@@ -4,10 +4,7 @@ import { LastLocationProvider } from 'react-router-last-location'
 import { useWeb3Context } from 'web3-react'
 
 import { ConnectedWeb3 } from '../../hooks/connectedWeb3'
-import { Header } from '../common/header'
-import { MainScroll } from '../common/main_scroll'
-import { MainWrapper } from '../common/main_wrapper'
-import { WrongNetworkMessage } from '../common/wrong_network_message'
+import { Header, MainScroll, MainWrapper, WrongNetworkMessage } from '../common'
 import { MarketHomeContainer } from '../market/market_home_container'
 import { MarketRoutes } from '../market/market_routes'
 import { MarketWizardCreatorContainer } from '../market/market_wizard_creator_container'
