+++ app/src/components/market/market_buy_container.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useConnectedWeb3Context } from '../../hooks/connectedWeb3'
 import { useMarketMakerData } from '../../hooks/useMarketMakerData'
 import { useQuestion } from '../../hooks/useQuestion'
-import { Loading } from '../common/loading'
+import { Loading } from '../common'
 
 import { MarketBuy } from './market_buy'
