+++ app/src/components/market/market_fund.tsx
@@ -13,22 +13,27 @@ import { ButtonType } from '../../theme/component_styles/button_styling_types'
 import { getLogger } from '../../util/logger'
 import { divBN, formatBigNumber } from '../../util/tools'
 import { BalanceItem, OutcomeTableValue, Status, Token } from '../../util/types'
-import { BalanceShares } from '../common/balance_shares'
-import { BalanceToken } from '../common/balance_token'
-import { BigNumberInput, BigNumberInputReturn } from '../common/big_number_input'
-import { Button } from '../common/button'
-import { ButtonContainer } from '../common/button_container'
-import { ButtonLink } from '../common/button_link'
-import { FormError } from '../common/form_error'
-import { FormLabel } from '../common/form_label'
-import { FormRow } from '../common/form_row'
-import { Loading } from '../common/loading'
-import { OutcomeTable } from '../common/outcome_table'
-import { SectionTitle } from '../common/section_title'
-import { SubsectionTitle } from '../common/subsection_title'
-import { TD, TR, Table } from '../common/table'
-import { TextfieldCustomPlaceholder } from '../common/textfield_custom_placeholder'
-import { ViewCard } from '../common/view_card/'
+import {
+  BalanceShares,
+  BalanceToken,
+  BigNumberInput,
+  Button,
+  ButtonContainer,
+  ButtonLink,
+  FormError,
+  FormLabel,
+  FormRow,
+  Loading,
+  OutcomeTable,
+  SectionTitle,
+  SubsectionTitle,
+  TD,
+  TR,
+  Table,
+  TextfieldCustomPlaceholder,
+  ViewCard,
+} from '../common'
+import { BigNumberInputReturn } from '../common/big_number_input'
 
 interface Props extends RouteComponentProps<any> {
   marketMakerAddress: string
