+++ app/src/components/market/market_sell.tsx
@@ -12,22 +12,28 @@ import { ButtonType } from '../../theme/component_styles/button_styling_types'
 import { getLogger } from '../../util/logger'
 import { calcSellAmountInCollateral, computeBalanceAfterTrade, formatBigNumber, mulBN } from '../../util/tools'
 import { BalanceItem, OutcomeTableValue, Status, Token } from '../../util/types'
-import { BigNumberInput, Button, OutcomeTable } from '../common'
-import { BalanceShares } from '../common/balance_shares'
+import {
+  BalanceShares,
+  BigNumberInput,
+  Button,
+  ButtonContainer,
+  ButtonLink,
+  FormError,
+  FormLabel,
+  FormRow,
+  Loading,
+  OutcomeTable,
+  Paragraph,
+  SectionTitle,
+  SubsectionTitle,
+  TD,
+  TR,
+  Table,
+  TextfieldCustomPlaceholder,
+  ViewCard,
+  Well,
+} from '../common'
 import { BigNumberInputReturn } from '../common/big_number_input'
-import { ButtonContainer } from '../common/button_container'
-import { ButtonLink } from '../common/button_link'
-import { FormError } from '../common/form_error'
-import { FormLabel } from '../common/form_label'
-import { FormRow } from '../common/form_row'
-import { Loading } from '../common/loading'
-import { Paragraph } from '../common/paragraph'
-import { SectionTitle } from '../common/section_title'
-import { SubsectionTitle } from '../common/subsection_title'
-import { TD, TR, Table } from '../common/table'
-import { TextfieldCustomPlaceholder } from '../common/textfield_custom_placeholder'
-import { ViewCard } from '../common/view_card'
-import { Well } from '../common/well'
 
 const ButtonLinkStyled = styled(ButtonLink)`
   margin-right: auto;
