+++ app/src/components/market/market_sell_container.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useConnectedWeb3Context } from '../../hooks/connectedWeb3'
 import { useMarketMakerData } from '../../hooks/useMarketMakerData'
 import { useQuestion } from '../../hooks/useQuestion'
-import { Loading } from '../common/loading'
+import { Loading } from '../common'
 
 import { MarketSell } from './market_sell'
